pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "1.0.0"
description = ""
authors = ["Stelios Giannikis <steliosgiannik@gmail.com>"]
readme = ""

[tool.poetry.dependencies]
python = "=3.10.12"
torch = "=2.2.1"
sentence-transformers = "=2.5.0"
uvicorn = "=0.27.1"
fastapi = "=0.110.0"
celery = {extras = ["redis"], version = "=5.3.6"}
redis = "=5.0.2"
python-dotenv = "=1.0.1"
flower = "=2.0.1"
psutil = "^6.0.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"